# BackEnd/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StoryX API",
    "version": "1.0.0",
    "description": "StoryX Backend API - AI-powered animation video generation platform",
    "contact": {
      "name": "StoryX Team"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Success message"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "Character": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "项目ID，如果为null则表示是全局角色（所有项目共享）"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "age": {
            "type": "string",
            "nullable": true,
            "description": "角色年龄，可能是\"25岁\"这样的描述"
          },
          "gender": {
            "type": "string"
          },
          "personality": {
            "type": "string",
            "description": "JSON array string"
          },
          "appearance": {
            "type": "string"
          },
          "style": {
            "type": "string"
          },
          "voiceActor": {
            "type": "string",
            "nullable": true,
            "description": "发音人"
          },
          "voiceTone": {
            "type": "string",
            "nullable": true,
            "description": "音色"
          },
          "voiceSample": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "音色示例（URL）"
          },
          "clothingStyle": {
            "type": "string",
            "nullable": true,
            "description": "服饰风格"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "description": "角色图片URL"
          },
          "videoUrl": {
            "type": "string",
            "nullable": true,
            "description": "角色视频URL（从抽卡任务生成）"
          },
          "modelUrl": {
            "type": "string",
            "nullable": true
          },
          "cached": {
            "type": "boolean"
          },
          "usageCount": {
            "type": "integer"
          },
          "project": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "title": {
                "type": "string"
              }
            },
            "description": "关联的项目信息（如果角色属于项目）"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/characters": {
      "get": {
        "summary": "获取当前用户的所有角色",
        "description": "获取当前登录用户的所有角色列表",
        "tags": [
          "Characters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cached",
            "schema": {
              "type": "boolean"
            },
            "description": "是否只返回已缓存的角色"
          },
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "项目ID，如果提供则只返回该项目和全局角色"
          },
          {
            "in": "query",
            "name": "novelId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "小说ID，如果提供则只返回该小说的角色（需要同时提供 projectId）"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回角色列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Character"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建新角色",
        "description": "创建一个新的角色",
        "tags": [
          "Characters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "小明"
                  },
                  "description": {
                    "type": "string",
                    "example": "角色描述"
                  },
                  "age": {
                    "type": "integer",
                    "example": 20
                  },
                  "gender": {
                    "type": "string",
                    "example": "male"
                  },
                  "personality": {
                    "type": "string",
                    "description": "性格特征JSON数组字符串",
                    "example": "[\"勇敢\", \"聪明\"]"
                  },
                  "appearance": {
                    "type": "string",
                    "example": "黑色短发，蓝色眼睛"
                  },
                  "style": {
                    "type": "string",
                    "example": "现代风格"
                  },
                  "voiceActor": {
                    "type": "string",
                    "description": "发音人",
                    "example": "张三"
                  },
                  "voiceTone": {
                    "type": "string",
                    "description": "音色",
                    "example": "温柔"
                  },
                  "voiceSample": {
                    "type": "string",
                    "format": "uri",
                    "description": "音色示例（URL）",
                    "example": "https://example.com/voice-sample.mp3"
                  },
                  "clothingStyle": {
                    "type": "string",
                    "description": "服饰风格",
                    "example": "现代休闲"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com/character.jpg"
                  },
                  "modelUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com/model.glb"
                  },
                  "projectId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "项目ID（可选），如果提供则角色属于该项目，否则为全局角色",
                    "example": "uuid-of-project"
                  },
                  "novelId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "小说ID（可选），如果提供则角色属于该小说（需要同时提供 projectId）",
                    "example": "uuid-of-novel"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "角色创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Character"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/{id}": {
      "get": {
        "summary": "获取角色详情",
        "description": "根据角色ID获取角色详细信息",
        "tags": [
          "Characters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "角色ID"
          },
          {
            "in": "query",
            "name": "novelId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "小说ID（可选，用于验证角色是否属于该小说）"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回角色详情",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Character"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "角色不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新角色",
        "description": "更新角色信息",
        "tags": [
          "Characters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "角色ID"
          },
          {
            "in": "query",
            "name": "novelId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "小说ID（可选，用于验证角色是否属于该小说）"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "personality": {
                    "type": "string"
                  },
                  "appearance": {
                    "type": "string"
                  },
                  "style": {
                    "type": "string"
                  },
                  "voiceActor": {
                    "type": "string",
                    "description": "发音人"
                  },
                  "voiceTone": {
                    "type": "string",
                    "description": "音色"
                  },
                  "voiceSample": {
                    "type": "string",
                    "format": "uri",
                    "description": "音色示例（URL）"
                  },
                  "clothingStyle": {
                    "type": "string",
                    "description": "服饰风格"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "modelUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "projectId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "项目ID（可选），如果提供则角色属于该项目，否则为全局角色"
                  },
                  "novelId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "小说ID（可选），如果提供则角色属于该小说（需要同时提供 projectId）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "角色更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Character"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "角色不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除角色",
        "description": "删除指定的角色",
        "tags": [
          "Characters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "角色ID"
          },
          {
            "in": "query",
            "name": "novelId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "小说ID（可选，用于验证角色是否属于该小说）"
          }
        ],
        "responses": {
          "200": {
            "description": "角色删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "description": "角色不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/merge": {
      "post": {
        "summary": "自动合并重复角色",
        "description": "合并同一用户、同一项目下同名的角色，保留最完整的信息，合并 shotIds。如果未提供 projectId，则合并用户下所有项目的重复角色。",
        "tags": [
          "Characters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "项目ID（可选），如果提供则只合并该项目下的角色，否则合并用户下所有项目的角色",
                    "example": "uuid-of-project"
                  },
                  "novelId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "小说ID（可选），如果提供则只合并该小说下的角色（需要同时提供 projectId）",
                    "example": "uuid-of-novel"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "合并成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "totalProcessed": {
                              "type": "integer",
                              "description": "处理的总角色数"
                            },
                            "mergedCount": {
                              "type": "integer",
                              "description": "合并的角色数（被删除的重复角色数）"
                            },
                            "deletedCount": {
                              "type": "integer",
                              "description": "删除的重复角色数"
                            },
                            "remainingCount": {
                              "type": "integer",
                              "description": "剩余的角色数（合并后）"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "项目不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{id}/characters": {
      "get": {
        "summary": "获取项目的所有角色",
        "description": "获取指定项目的所有角色（包括项目专属角色和全局角色）",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回角色列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Character"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "项目不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/novels/{novelId}/acts": {
      "get": {
        "summary": "获取小说的所有剧幕（按章节顺序排序）",
        "description": "获取指定小说的所有剧幕，按照章节顺序排序，确保即使异步处理也不会打乱顺序",
        "tags": [
          "Script"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "novelId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "小说ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回剧幕列表（按章节顺序排序）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Act"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "小说不存在"
          }
        }
      }
    },
...
  },
...
}